move classic editor shortcodes plugin to plugin folder
