+++ packages/js/src/classic-editor/plugins/shortcodes.js
